+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- None
+- Fixed: [pymatcalc.py] examples
 
 ### Added
 
-- None
+- Added Pester tests and GitHub workflows
 
 
 ## [1.2.0]