+++ CHANGELOG.md
@@ -9,8 +9,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- None
+
+### Added
+
+- None
+
+
+## [1.2.1]
+
+### Changed
+
 - Fixed delimiter to use raw string [catcsv.py], [pycalc.py], [pymatcalc.py], [pyplot-pandas.py], [pyplot-timeline2.py], [pyplot-x-rs.py], [pyplot.py], [pysym.py]
 - Fixed examples [pymatcalc.py] 
+- Updated [README.md]
 
 ### Added
 
@@ -64,7 +76,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [CHANGELOG.md]: blob/main/CHANGELOG.md
 
 
-[unreleased]: https://github.com/btklab/py-mocks/compare/1.2.0..HEAD
+[unreleased]: https://github.com/btklab/py-mocks/compare/1.2.1..HEAD
+[1.2.1]: https://github.com/btklab/py-mocks/releases/tag/1.2.1
 [1.2.0]: https://github.com/btklab/py-mocks/releases/tag/1.2.0
 [1.1.0]: https://github.com/btklab/py-mocks/releases/tag/1.1.0
 [1.0.0]: https://github.com/btklab/py-mocks/releases/tag/1.0.0